v3.2.3数据库-自动清理多余数据

+++ netspylib/src/main/java/com/creditease/netspy/inner/ui/bugspy/BugSpyListFragment.java
@@ -4,6 +4,7 @@ import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.DividerItemDecoration;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -18,11 +19,11 @@ import android.widget.Toast;
 import com.creditease.netspy.DBHelper;
 import com.creditease.netspy.R;
 import com.creditease.netspy.inner.db.BugEvent;
-import com.creditease.netspy.inner.db.HttpEvent;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * Created by zhxh on 2019/07/16
@@ -65,10 +66,16 @@ public class BugSpyListFragment extends Fragment {
     public void updateDataFromDb() {
         List<BugEvent> dataList = DBHelper.getInstance().getAllBugData();
         if (dataList.size() > 200) {
-            Toast.makeText(getActivity(), "异常崩溃数据已经达到" + dataList.size() + "条，为防止数据过多已经自动清理", Toast.LENGTH_LONG).show();
-            adapter.setData(new ArrayList<>());
-            DBHelper.getInstance().deleteAllBugData();
-            return;
+            AlertDialog dialog = new AlertDialog.Builder(Objects.requireNonNull(getContext()))
+                    .setTitle("温馨提示")
+                    .setMessage("异常崩溃数据已经达到" + dataList.size() + "条，为防止数据过多请自动清理")
+                    .setPositiveButton("清理", (dialog1, which) -> {
+                        adapter.setData(new ArrayList<>());
+                        DBHelper.getInstance().deleteAllBugData();
+                    })
+                    .setNegativeButton("取消", null)
+                    .create();
+            dialog.show();
         } else if (dataList.size() > 100) {
             Toast.makeText(getActivity(), "异常崩溃数据已经达到" + dataList.size() + "条，请按主动屏幕右上角删除按钮及时清理（当数据超过200条会触发自动清理）", Toast.LENGTH_LONG).show();
         }
